backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# AI & LLM - Let pip resolve versions
langgraph
langchain
langchain-core
langchain-community
google-generativeai
langfuse

# File Processing
python-multipart==0.0.6
Pillow==10.2.0
PyPDF2==3.0.1
pdf2image==1.16.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2

# Security
cryptography==42.0.0
bcrypt==4.1.2
